refactor: Restructure user service and update Makefile for environment reset
